# api/openapi.yaml
openapi: 3.0.3
info:
  title: SOJ v2 API
  version: 0.1.0
  description: SOJ v2 REST contract baseline. All 2xx responses use the response envelope except 204 and explicitly empty 202 responses.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/contests:
    get:
      tags: [Contests]
      operationId: listContests
      summary: List contests
      x-owner: WP5 Contest
      x-auth: public
      security: []
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ContestStatus" }
        - name: visibility
          in: query
          schema: { $ref: "#/components/schemas/ContestVisibility" }
        - name: keyword
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Contest page.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContestPageEnvelope"
        "400": { $ref: "#/components/responses/BadRequest" }
        "500": { $ref: "#/components/responses/InternalError" }
    post:
      tags: [Contests]
      operationId: createContest
      summary: Create a contest
      x-owner: WP5 Contest
      x-auth: user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContestCreateRequest"
      responses:
        "201":
          description: Contest created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContestEnvelope"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PermissionDenied" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/contests/{id}:
    parameters:
      - $ref: "#/components/parameters/ID"
    get:
      tags: [Contests]
      operationId: getContest
      summary: Get a contest
      x-owner: WP5 Contest
      x-auth: public-conditional
      security: []
      responses:
        "200":
          description: Contest.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContestEnvelope"
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PermissionDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalError" }
    patch:
      tags: [Contests]
      operationId: updateContest
      summary: Update a contest
      x-owner: WP5 Contest
      x-auth: owner-admin-root
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContestUpdateRequest"
      responses:
        "200":
          description: Contest updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContestEnvelope"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PermissionDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "500": { $ref: "#/components/responses/InternalError" }
    delete:
      tags: [Contests]
      operationId: deleteContest
      summary: Delete a contest
      x-owner: WP5 Contest
      x-auth: owner-admin-root
      responses:
        "204":
          description: Contest archived/deleted. Empty body.
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403": { $ref: "#/components/responses/PermissionDenied" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalError" }
  /api/v1/contests/cursor:
    get:
      tags: [Contests]
      operationId: listContestsByCursor
      summary: List contests with a cursor
      x-owner: WP5 Contest
      x-auth: public
      security: []
      parameters:
        - $ref: "#/components/parameters/PageSize"
        - $ref: "#/components/parameters/Cursor"
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ContestStatus" }
        - name: visibility
          in: query
          schema: { $ref: "#/components/schemas/ContestVisibility" }
        - name: keyword
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Cursor page of contests without a total count.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContestCursorPageEnvelope"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "500": { $ref: "#/components/responses/InternalError" }
...
components:
  securitySchemes:
    bearerAuth:
      x-owner: WP1 Data/API Contract
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ID:
      x-owner: WP1 Data/API Contract
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
    Page:
      x-owner: WP1 Data/API Contract
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      x-owner: WP1 Data/API Contract
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      x-owner: WP1 Data/API Contract
      name: cursor
      in: query
      description: Opaque cursor returned as next_cursor by the previous response.
      schema: { type: string }
  responses:
    BadRequest:
      x-owner: WP1 Data/API Contract
      description: Invalid argument.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    Unauthenticated:
      x-owner: WP1 Data/API Contract
      description: Missing or invalid authentication.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    PermissionDenied:
      x-owner: WP1 Data/API Contract
      description: Permission denied.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    NotFound:
      x-owner: WP1 Data/API Contract
      description: Resource not found.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    Conflict:
      x-owner: WP1 Data/API Contract
      description: Resource conflict.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
    ValidationFailed:
      x-owner: WP1 Data/API Contract
      description: Validation failed.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
...
    InternalError:
      x-owner: WP1 Data/API Contract
      description: Internal error.
      content: { application/json: { schema: { $ref: "#/components/schemas/ErrorEnvelope" } } }
  schemas:
    Envelope:
      x-owner: WP1 Data/API Contract
      type: object
      required: [data, error, request_id]
      properties:
        data: {}
        error:
          nullable: true
          $ref: "#/components/schemas/Error"
        request_id:
          type: string
    ErrorEnvelope:
      x-owner: WP1 Data/API Contract
      allOf:
        - $ref: "#/components/schemas/Envelope"
        - type: object
          properties:
            data:
              type: object
              nullable: true
            error:
              $ref: "#/components/schemas/Error"
    Error:
      x-owner: WP1 Data/API Contract
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          enum:
            - invalid_argument
            - unauthenticated
            - permission.denied
            - route.not_found
            - validation.failed
            - auth.invalid_credentials
            - auth.required
            - auth.token_expired
            - auth.refresh_token_revoked
            - user.email_exists
            - problem.id_invalid
            - problem.forbidden
            - problem.not_found
            - problem.slug_conflict
            - problem.not_publishable
            - problem.testcase_not_ready
            - problem_check.id_invalid
            - problem_check.not_found
            - testcase.invalid_archive
            - submission.not_found
            - submission.not_allowed
            - submission.language_disabled
            - run.not_found
            - contest.not_found
            - contest.not_started
            - contest.ended
            - contest.registration_required
            - contest.scoreboard_hidden
            - queue.dispatch_failed
            - judge.engine_unavailable
            - service_unavailable
            - internal_error
        message:
          type: string
    PageMeta:
      x-owner: WP1 Data/API Contract
      type: object
      required: [page, page_size, total]
      properties:
        page: { type: integer, minimum: 1 }
        page_size: { type: integer, minimum: 1 }
        total: { type: integer, format: int64, minimum: 0 }
...
    ContestVisibility:
      type: string
      enum: [public, private]
    ContestStatus:
      type: string
      enum: [draft, published, running, ended, archived]
    ContestScoringMode:
      type: string
      description: Contest scoring mode. Current backend support is ACM.
      enum: [acm]
...
    ContestProblemInput:
      type: object
      required: [problem_id, alias]
      properties:
        problem_id: { type: integer, format: int64 }
        alias: { type: string }
    ContestProblem:
      type: object
      required: [problem_id, alias, sort_order]
      properties:
        problem_id: { type: integer, format: int64 }
        alias: { type: string }
        sort_order: { type: integer }
        title:
          type: string
          description: Problem title enriched from the linked problem when available.
    ContestCreateRequest:
      type: object
      required: [title, visibility, start_at, end_at, freeze_at, problems]
      properties:
        title: { type: string }
        description: { type: string }
        visibility: { $ref: "#/components/schemas/ContestVisibility" }
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        freeze_at: { type: string, format: date-time }
        invite_code: { type: string, description: Required when visibility is private. }
        problems:
          type: array
          items: { $ref: "#/components/schemas/ContestProblemInput" }
    ContestUpdateRequest:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        visibility: { $ref: "#/components/schemas/ContestVisibility" }
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        freeze_at: { type: string, format: date-time }
        invite_code: { type: string, description: Required when changing visibility to private without an existing invite code. }
        problems:
          type: array
          items: { $ref: "#/components/schemas/ContestProblemInput" }
        status: { $ref: "#/components/schemas/ContestStatus" }
    ContestResponse:
      type: object
      required: [id, owner_user_id, title, visibility, status, scoring_mode, registered, start_at, end_at, freeze_at, problems, created_at, updated_at]
      properties:
        id: { type: integer, format: int64 }
        owner_user_id: { type: integer, format: int64 }
        title: { type: string }
        description: { type: string, nullable: true }
        visibility: { $ref: "#/components/schemas/ContestVisibility" }
        status: { $ref: "#/components/schemas/ContestStatus" }
        scoring_mode: { $ref: "#/components/schemas/ContestScoringMode" }
        registered:
          type: boolean
          description: Whether the current authenticated user has an active registration for this contest. Anonymous users receive false.
        start_at: { type: string, format: date-time }
        end_at: { type: string, format: date-time }
        freeze_at: { type: string, format: date-time }
        problems:
          type: array
          items: { $ref: "#/components/schemas/ContestProblem" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ContestPage:
      allOf:
        - $ref: "#/components/schemas/PageMeta"
        - type: object
          required: [items]
          properties:
            items: { type: array, items: { $ref: "#/components/schemas/ContestResponse" } }
    ContestCursorPage:
      type: object
      required: [items]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/ContestResponse" } }
        next_cursor:
          type: string
          description: Opaque cursor for the next page. Omitted when no more items remain.
...
    ContestEnvelope:
      allOf: [{ $ref: "#/components/schemas/Envelope" }]
      properties: { data: { $ref: "#/components/schemas/ContestResponse" } }
    ContestPageEnvelope:
      allOf: [{ $ref: "#/components/schemas/Envelope" }]
      properties: { data: { $ref: "#/components/schemas/ContestPage" } }
    ContestCursorPageEnvelope:
      allOf: [{ $ref: "#/components/schemas/Envelope" }]
      properties: { data: { $ref: "#/components/schemas/ContestCursorPage" } }
...